Crate numeric

Numeric Rust provides a foundation for doing scientific computing with Rust. It aims to be for Rust what Numpy is for Python.

OpenBLAS/LAPACK is used to make things like matrix muliplications and solving linear equations fast.

Reexports

pub use tensor::{Tensor, AxisIndex};

pub use tensor::{SingleTensor, DoubleTensor};

pub use math::{log, ln, log10, log2, sin, cos, tan, asin, acos, atan, exp_m1, exp, exp2, ln_1p, sinh, cosh, tanh, asinh, acosh, atanh, atan2, sqrt, floor, ceil, round, trunc, fract, abs, signum, powf, powi, is_nan, is_finite, is_infinite, is_normal, is_sign_positive, is_sign_negative};

Modules

io
Saving and loading data to and from disk.

linalg
Linear algebra functions.

math
Contains mathematical functions that operate on tensors. These functions are largely modelled after what is available natively in Rust.

random
The random module provides methods of randomizing tensors.

tensor
The tensor module defines an N-dimensional matrix for use in scientific computing.

Macros

tensor!
Macro for creating vectors and matrices.

Traits

Numeric
Numeric extends TensorType to all the numeric types supported by Tensor (e.g. u8 and f32).

TensorType
This is the basic trait that must be satisfied for basic elements used in Tensor.