Module numeric::tensor

The tensor module defines an N-dimensional matrix for use in scientific computing.

Many of the things in this module are lifted out of the tensor namespace, which means you can do things like:

use numeric::Tensor;

Structs

Tensor
An implementation of an N-dimensional matrix. A quick example:

Enums

AxisIndex
Used for advanced slicing of a Tensor.

Type Definitions

DoubleTensor
Type alias for Tensor<f64>

SingleTensor
Type alias for Tensor<f32>